+++ tests/Kiota.Builder.Tests/Extensions/StringExtensionsTests.cs
@@ -39,6 +39,23 @@ namespace Kiota.Builder.Extensions.Tests {
             Assert.Equal(string.Empty, string.Empty.ReplaceValueIdentifier());
             Assert.Equal("microsoft.graph.message.Content", "microsoft.graph.message.$value".ReplaceValueIdentifier());
         }
+        [Fact]
+        public void ToSnakeCase() {
+            string nString = null;
+            Assert.Null(nString.ToSnakeCase());
+            Assert.Equal(string.Empty, string.Empty.ToSnakeCase());
+            Assert.Equal("toto", "Toto".ToSnakeCase());
+            Assert.Equal("microsoft_graph_message_content", "microsoft-Graph-Message-Content".ToSnakeCase());
+            Assert.Equal("microsoft_graph_message_content", "microsoftGraphMessageContent".ToSnakeCase());
+        }
+        [Fact]
+        public void NormalizeNameSpaceName() {
+            string nString = null;
+            Assert.Null(nString.NormalizeNameSpaceName("."));
+            Assert.Equal(string.Empty, string.Empty.NormalizeNameSpaceName("."));
+            Assert.Equal("Toto", "toto".NormalizeNameSpaceName("-"));
+            Assert.Equal("Microsoft_Graph_Message_Content", "microsoft.Graph.Message.Content".NormalizeNameSpaceName("_"));
+        }
     }
     
 }