+++ tests/Kiota.Builder.Tests/Writers/LanguageWriterTests.cs
@@ -2,6 +2,7 @@ using System.ComponentModel;
 using System.IO;
 using Kiota.Builder.Writers.CSharp;
 using Kiota.Builder.Writers.Java;
+using Kiota.Builder.Writers.Ruby;
 using Kiota.Builder.Writers.TypeScript;
 using Xunit;
 
@@ -15,6 +16,8 @@ namespace Kiota.Builder.Writers.Tests {
                         LanguageWriter.GetLanguageWriter(GenerationLanguage.CSharp, defaultPath, defaultName).GetType());
             Assert.Equal(typeof(JavaWriter),
                         LanguageWriter.GetLanguageWriter(GenerationLanguage.Java, defaultPath, defaultName).GetType());
+            Assert.Equal(typeof(RubyWriter),
+                        LanguageWriter.GetLanguageWriter(GenerationLanguage.Ruby, defaultPath, defaultName).GetType());
             Assert.Equal(typeof(TypeScriptWriter),
                         LanguageWriter.GetLanguageWriter(GenerationLanguage.TypeScript, defaultPath, defaultName).GetType());
             Assert.Throws<InvalidEnumArgumentException>(() => LanguageWriter.GetLanguageWriter(GenerationLanguage.Go, defaultPath, defaultName));