+++ tests/Kiota.Builder.Tests/Writers/Ruby/CodeMethodWriterTests.cs
@@ -1,0 +1,308 @@
+using System;
+using System.IO;
+using System.Linq;
+using Kiota.Builder.Extensions;
+using Kiota.Builder.Tests;
+using Xunit;
+
+namespace Kiota.Builder.Writers.Ruby.Tests {
+    public class CodeMethodWriterTests : IDisposable {
+        private const string defaultPath = "./";
+        private const string defaultName = "name";
+        private readonly StringWriter tw;
+        private readonly LanguageWriter writer;
+        private readonly CodeMethod method;
+        private readonly CodeClass parentClass;
+        private const string methodName = "methodName";
+        private const string returnTypeName = "Somecustomtype";
+        private const string methodDescription = "some description";
+        private const string paramDescription = "some parameter description";
+        private const string paramName = "paramName";
+        public CodeMethodWriterTests()
+        {
+            writer = LanguageWriter.GetLanguageWriter(GenerationLanguage.Ruby, defaultPath, defaultName);
+            tw = new StringWriter();
+            writer.SetTextWriter(tw);
+            var root = CodeNamespace.InitRootNamespace();
+            parentClass = new CodeClass(root) {
+                Name = "parentClass"
+            };
+            root.AddClass(parentClass);
+            method = new CodeMethod(parentClass) {
+                Name = methodName,
+            };
+            method.ReturnType = new CodeType(method) {
+                Name = returnTypeName
+            };
+            parentClass.AddMethod(method);
+        }
+        public void Dispose()
+        {
+            tw?.Dispose();
+        }
+        private void AddSerializationProperties() {
+            var addData = parentClass.AddProperty(new CodeProperty(parentClass) {
+                Name = "additionalData",
+                PropertyKind = CodePropertyKind.AdditionalData,
+            }).First();
+            addData.Type = new CodeType(addData) {
+                Name = "string"
+            };
+            var dummyProp = parentClass.AddProperty(new CodeProperty(parentClass) {
+                Name = "dummyProp",
+            }).First();
+            dummyProp.Type = new CodeType(dummyProp) {
+                Name = "string"
+            };
+            var dummyCollectionProp = parentClass.AddProperty(new CodeProperty(parentClass) {
+                Name = "dummyColl",
+            }).First();
+            dummyCollectionProp.Type = new CodeType(dummyCollectionProp) {
+                Name = "string",
+                CollectionKind = CodeTypeBase.CodeTypeCollectionKind.Array,
+            };
+            var dummyComplexCollection = parentClass.AddProperty(new CodeProperty(parentClass) {
+                Name = "dummyComplexColl"
+            }).First();
+            dummyComplexCollection.Type = new CodeType(dummyComplexCollection) {
+                Name = "Complex",
+                CollectionKind = CodeTypeBase.CodeTypeCollectionKind.Array,
+                TypeDefinition = new CodeClass(parentClass.Parent) {
+                    Name = "SomeComplexType"
+                }
+            };
+            var dummyEnumProp = parentClass.AddProperty(new CodeProperty(parentClass){
+                Name = "dummyEnumCollection",
+            }).First();
+            dummyEnumProp.Type = new CodeType(dummyEnumProp) {
+                Name = "SomeEnum",
+                TypeDefinition = new CodeEnum(parentClass.Parent) {
+                    Name = "EnumType"
+                }
+            };
+        }
+        private void AddInheritanceClass() {
+            (parentClass.StartBlock as CodeClass.Declaration).Inherits = new CodeType(parentClass) {
+                Name = "someParentClass"
+            };
+        }
+        private void AddRequestBodyParameters() {
+            var stringType = new CodeType(method) {
+                Name = "string",
+            };
+            method.AddParameter(new CodeParameter(method) {
+                Name = "h",
+                ParameterKind = CodeParameterKind.Headers,
+                Type = stringType,
+            });
+            method.AddParameter(new CodeParameter(method){
+                Name = "q",
+                ParameterKind = CodeParameterKind.QueryParameter,
+                Type = stringType,
+            });
+            method.AddParameter(new CodeParameter(method){
+                Name = "b",
+                ParameterKind = CodeParameterKind.RequestBody,
+                Type = stringType,
+            });
+            method.AddParameter(new CodeParameter(method){
+                Name = "r",
+                ParameterKind = CodeParameterKind.ResponseHandler,
+                Type = stringType,
+            });
+        }
+        [Fact]
+        public void WritesRequestBodiesThrowOnNullHttpMethod() {
+            method.MethodKind = CodeMethodKind.RequestExecutor;
+            Assert.Throws<InvalidOperationException>(() => writer.Write(method));
+            method.MethodKind = CodeMethodKind.RequestGenerator;
+            Assert.Throws<InvalidOperationException>(() => writer.Write(method));
+        }
+        [Fact]
+        public void WritesRequestExecutorBody() {
+            method.MethodKind = CodeMethodKind.RequestExecutor;
+            method.HttpMethod = HttpMethod.Get;
+            AddRequestBodyParameters();
+            writer.Write(method);
+            var result = tw.ToString();
+            Assert.Contains("request_info", result);
+            Assert.Contains("send_async", result);
+            AssertExtensions.CurlyBracesAreClosed(result);
+        }
+        [Fact]
+        public void WritesRequestGeneratorBody() {
+            method.MethodKind = CodeMethodKind.RequestGenerator;
+            method.HttpMethod = HttpMethod.Get;
+            AddRequestBodyParameters();
+            writer.Write(method);
+            var result = tw.ToString();
+            Assert.Contains("request_info = RequestInfo.new", result);
+            Assert.Contains("http_method = :GET", result);
+            Assert.Contains("set_query_string_parameters_from_raw_object", result);
+            Assert.Contains("set_content_from_parsable", result);
+            Assert.Contains("return request_info;", result);
+        }
+        [Fact]
+        public void WritesDeSerializerBody() {
+            var parameter = new CodeParameter(method){
+                Description = paramDescription,
+                Name = paramName
+            };
+            parameter.Type = new CodeType(parameter) {
+                Name = "string"
+            };
+            method.MethodKind = CodeMethodKind.Deserializer;
+            method.IsAsync = false;
+            AddSerializationProperties();
+            writer.Write(method);
+            var result = tw.ToString();
+            Assert.Contains("get_collection_of_primitive_values", result);
+            Assert.Contains("get_collection_of_object_values", result);
+            Assert.Contains("get_enum_value", result);
+        }
+        [Fact]
+        public void WritesInheritedSerializerBody() {
+            method.MethodKind = CodeMethodKind.Serializer;
+            method.IsAsync = false;
+            AddSerializationProperties();
+            AddInheritanceClass();
+            writer.Write(method);
+            var result = tw.ToString();
+            Assert.Contains("super.serialize", result);
+            AssertExtensions.CurlyBracesAreClosed(result);
+        }
+        [Fact]
+        public void WritesSerializerBody() {
+            var parameter = new CodeParameter(method){
+                Description = paramDescription,
+                Name = paramName
+            };
+            parameter.Type = new CodeType(parameter) {
+                Name = "string"
+            };
+            method.MethodKind = CodeMethodKind.Serializer;
+            method.IsAsync = false;
+            AddSerializationProperties();
+            writer.Write(method);
+            var result = tw.ToString();
+            Assert.Contains("write_collection_of_primitive_values", result);
+            Assert.Contains("write_collection_of_object_values", result);
+            Assert.Contains("write_enum_value", result);
+            Assert.Contains("write_additional_data", result);
+            AssertExtensions.CurlyBracesAreClosed(result);
+        }
+        [Fact]
+        public void WritesMethodSyncDescription() {
+            
+            method.Description = methodDescription;
+            method.IsAsync = false;
+            var parameter = new CodeParameter(method){
+                Description = paramDescription,
+                Name = paramName
+            };
+            parameter.Type = new CodeType(parameter) {
+                Name = "string"
+            };
+            method.AddParameter(parameter);
+            writer.Write(method);
+            var result = tw.ToString();
+            Assert.DoesNotContain("@return a CompletableFuture of", result);
+            AssertExtensions.CurlyBracesAreClosed(result);
+        }
+        [Fact]
+        public void Defensive() {
+            var codeMethodWriter = new CodeMethodWriter(new RubyConventionService());
+            Assert.Throws<ArgumentNullException>(() => codeMethodWriter.WriteCodeElement(null, writer));
+            Assert.Throws<ArgumentNullException>(() => codeMethodWriter.WriteCodeElement(method, null));
+            var originalParent = method.Parent;
+            method.Parent = CodeNamespace.InitRootNamespace();
+            Assert.Throws<InvalidOperationException>(() => codeMethodWriter.WriteCodeElement(method, writer));
+        }
+        [Fact]
+        public void ThrowsIfParentIsNotClass() {
+            method.Parent = CodeNamespace.InitRootNamespace();
+            Assert.Throws<InvalidOperationException>(() => writer.Write(method));
+        }
+        private const string taskPrefix = "CompletableFuture<";
+        [Fact]
+        public void DoesNotAddAsyncInformationOnSyncMethods() {
+            method.IsAsync = false;
+            writer.Write(method);
+            var result = tw.ToString();
+            Assert.DoesNotContain(taskPrefix, result);
+            AssertExtensions.CurlyBracesAreClosed(result);
+        }
+        [Fact]
+        public void WritesGetterToField() {
+            method.AddAccessedProperty();
+            method.MethodKind = CodeMethodKind.Getter;
+            writer.Write(method);
+            var result = tw.ToString();
+            Assert.Contains("@some_property", result);
+        }
+        [Fact]
+        public void WritesSetterToField() {
+            method.AddAccessedProperty();
+            method.MethodKind = CodeMethodKind.Setter;
+            writer.Write(method);
+            var result = tw.ToString();
+            Assert.Contains("@some_property =", result);
+        }
+        [Fact]
+        public void WritesConstructor() {
+            method.MethodKind = CodeMethodKind.Constructor;
+            var defaultValue = "someval";
+            var propName = "propWithDefaultValue";
+            parentClass.AddProperty(new CodeProperty(parentClass) {
+                Name = propName,
+                DefaultValue = defaultValue,
+                PropertyKind = CodePropertyKind.PathSegment,
+            });
+            writer.Write(method);
+            var result = tw.ToString();
+            Assert.Contains($"@{propName.ToSnakeCase()} = {defaultValue}", result);
+        }
+        [Fact]
+        public void WritesApiConstructor() {
+            method.MethodKind = CodeMethodKind.ClientConstructor;
+            var coreProp = parentClass.AddProperty(new CodeProperty(parentClass) {
+                Name = "core",
+                PropertyKind = CodePropertyKind.HttpCore,
+            }).First();
+            coreProp.Type = new CodeType(coreProp) {
+                Name = "HttpCore",
+                IsExternal = true,
+            };
+            method.AddParameter(new CodeParameter(method) {
+                Name = "core",
+                ParameterKind = CodeParameterKind.HttpCore,
+                Type = coreProp.Type,
+            });
+            writer.Write(method);
+            var result = tw.ToString();
+            Assert.Contains(coreProp.Name, result);
+        }
+        [Fact]
+        public void WritesApiConstructorWithBackingStore() {
+            method.MethodKind = CodeMethodKind.ClientConstructor;
+            var coreProp = parentClass.AddProperty(new CodeProperty(parentClass) {
+                Name = "core",
+                PropertyKind = CodePropertyKind.HttpCore,
+            }).First();
+            coreProp.Type = new CodeType(coreProp) {
+                Name = "HttpCore",
+                IsExternal = true,
+            };
+            method.AddParameter(new CodeParameter(method) {
+                Name = "core",
+                ParameterKind = CodeParameterKind.HttpCore,
+                Type = coreProp.Type,
+            });
+            var tempWriter = LanguageWriter.GetLanguageWriter(GenerationLanguage.Java, defaultPath, defaultName, true);
+            tempWriter.SetTextWriter(tw);
+            tempWriter.Write(method);
+            var result = tw.ToString();
+            Assert.Contains("enableBackingStore", result);
+        }
+    }
+}