+++ tests/Kiota.Builder.Tests/Writers/Ruby/CodePropertyWriterTests.cs
@@ -1,0 +1,55 @@
+using System;
+using System.IO;
+using System.Linq;
+using Kiota.Builder.Extensions;
+using Xunit;
+
+namespace Kiota.Builder.Writers.Ruby.Tests {
+    public class CodePropertyWriterTests: IDisposable {
+        private const string defaultPath = "./";
+        private const string defaultName = "name";
+        private readonly StringWriter tw;
+        private readonly LanguageWriter writer;
+        private readonly CodeProperty property;
+        private readonly CodeClass parentClass;
+        private const string propertyName = "propertyName";
+        private const string propertyDescription = "some description";
+        private const string typeName = "Somecustomtype";
+        public CodePropertyWriterTests() {
+            writer = LanguageWriter.GetLanguageWriter(GenerationLanguage.Ruby, defaultPath, defaultName);
+            tw = new StringWriter();
+            writer.SetTextWriter(tw);
+            var root = CodeNamespace.InitRootNamespace();
+            parentClass = new CodeClass(root) {
+                Name = "parentClass"
+            };
+            root.AddClass(parentClass);
+            property = new CodeProperty(parentClass) {
+                Name = propertyName,
+            };
+            property.Type = new CodeType(property) {
+                Name = typeName
+            };
+            parentClass.AddProperty(property);
+        }
+        public void Dispose() {
+            tw?.Dispose();
+        }
+        [Fact]
+        public void WritesRequestBuilder() {property.PropertyKind = CodePropertyKind.RequestBuilder;
+            writer.Write(property);
+            var result = tw.ToString();
+            Assert.Contains($"def {propertyName.ToSnakeCase()}", result);
+            Assert.Contains($"{typeName}.new", result);
+            Assert.Contains("http_core", result);
+            Assert.Contains("path_segment", result);
+        }
+        [Fact]
+        public void WritesCustomProperty() {
+            property.PropertyKind = CodePropertyKind.Custom;
+            writer.Write(property);
+            var result = tw.ToString();
+            Assert.Contains($"@{propertyName.ToSnakeCase()}", result);
+        }
+    }
+}